refactor(price): implement thousandth cast and update price items handling
Introduce ThousandthCast for precise decimal handling in price items and improve data type consistency:
- Add new ThousandthCast class for handling thousandths precision
- Update MoneyCast spacing for consistency
- Modify PriceItem model to use ThousandthCast
- Change time_item field type to bigInteger for larger number support
- Update PriceItemsRelationManager to handle floats instead of integers
- Improve price name generation in factory

BREAKING CHANGE: Database schema changed for price_items.time_item column from integer to bigInteger

+++ app/Filament/Resources/PriceResource/RelationManagers/PriceItemsRelationManager.php
@@ -26,12 +26,12 @@ class PriceItemsRelationManager extends RelationManager
                     ->live(onBlur: true)
                     ->afterStateHydrated(function (?string $state, Get $get, Set $set) {
                         if ($state) {
-                            $set('time_item', intval($state));
+                            $set('time_item', floatval($state));
                         }
                     })
                     ->afterStateUpdated(function (?string $state, Get $get, Set $set) {
                         if ($state) {
-                            $set('time_item', intval($state));
+                            $set('time_item', floatval($state));
                         }
                     }),
                 Forms\Components\TextInput::make('time_item')
